Fix PDF printer install
The cups-pdf printer scriptlet on install wasn't working. Needed to manually run the lpadmin command and toggle cupsenable/cupsdisable/cupsenable to get the /etc/cups/printers.conf file to populate. This works fine manually in a WSL terminal, but requires a few delays to get to work when running from PowerShell. Removed cups-pdf reinstall section as it is now no longer needed.

- Update also includes RHEL->Alma label changes and removal of Oracle Linux mentions, as WSL Oracle is not going to be supported as it requires a lot more setup.